move `prefixUtilityClasses` helper to Platform
