+++ src/applications/personalization/profile-2/components/direct-deposit/DirectDepositContent.jsx
@@ -29,7 +29,7 @@ import ProfileInfoTable from '../ProfileInfoTable';
 import FraudVictimAlert from './FraudVictimAlert';
 import AdditionalInformation from './DirectDepositInformation';
 
-import { prefixUtilityClasses } from '../../helpers';
+import prefixUtilityClasses from 'platform/utilities/prefix-utility-classes';
 
 export const DirectDepositContent = ({
   isAuthorized,
