+++ src/applications/personalization/profile-2/tests/helpers.unit.spec.js
@@ -4,7 +4,6 @@ import { expect } from 'chai';
 
 import {
   getServiceBranchDisplayName,
-  prefixUtilityClasses,
   transformServiceHistoryEntryIntoTableRow,
 } from '../helpers';
 
@@ -24,25 +23,6 @@ describe('getServiceBranchDisplayName', () => {
   describe('when not passed one of the five USA military branches', () => {
     it('should simply return what it was passed', () => {
       expect(getServiceBranchDisplayName('Other')).to.equal('Other');
-    });
-  });
-});
-
-describe('prefixUtilityClasses', () => {
-  const classes = ['class-1', 'class-2'];
-  it('should prefix an array of classes with `vads-u-`', () => {
-    const expectedResult = ['vads-u-class-1', 'vads-u-class-2'];
-    const result = prefixUtilityClasses(classes);
-    expect(result).to.deep.equal(expectedResult);
-  });
-  describe('when passed a screenSize', () => {
-    it('should prefix an array of classes with the responsive prefix and `vads-u-`', () => {
-      const expectedResult = [
-        'medium-screen:vads-u-class-1',
-        'medium-screen:vads-u-class-2',
-      ];
-      const result = prefixUtilityClasses(classes, 'medium');
-      expect(result).to.deep.equal(expectedResult);
     });
   });
 });
